Some adjusts to use less memory.

+++ play/play.ino
@@ -1,4 +1,4 @@
-#include "song.h" //Put the name of your song file here
+#include "output.h" //Put the name of your song file here
 
 #define fstLed 6
 #define sndLed 5
@@ -13,9 +13,9 @@ void setup() {
   pinMode(8, OUTPUT);//Buzzer
   pinMode(fstLed, OUTPUT);//Led indicator when singing a note
   pinMode(sndLed, OUTPUT);//Led indicator when singing a note
-
+  
   //If you want to repeat the song, just put the following lines into the loop function
-  for(int i = 0; i < NumNotas; i++) {
+  for(short int i = 0; i < NumNotas; i++) {
     play(i, 2);
   }
 }
@@ -24,7 +24,7 @@ void loop() {
   //No need to repeat the song
 }
 
-void play(int i, int numberLedsUse) {
+void play(short int i, short int numberLedsUse) {
   int ledOn = fstLed;
   switch (numberLedsUse) {
     case 1:
@@ -43,9 +43,10 @@ void play(int i, int numberLedsUse) {
   }
 }
 
-void buzzer(int note, int duration) {
+void buzzer(short int note, short int duration) {
   tone(8, note, duration);
   delay(duration);
   noTone(8);
 }
+
